requirements.txt
flask>=3.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0
pyyaml>=6.0
rich>=13.7.0
plotly>=5.18.0
pytest>=8.0.0
requests>=2.31.0
